Add J1939 Address Claiming test functions
This patch introduces a new file, tst-j1939-ac.c, which includes a
series of test functions designed to evaluate the kernel side helpers
for the J1939 Address Claiming support. The tests aim to verify the
correct functioning of the J1939/ISOBUS traffic monitoring, the internal
cache of NAMEs with address mapping, as well as the resolution of NAMEs
for ingress and egress packets.

In addition, the tests evaluate the behavior of kernel J1939 stack when
applying certain restrictions, such as enforcing a 250ms pause between
address claimed messages and the rest of the communication. This helps
ensure that the kernel side helpers are correctly handling the J1939
Address Claiming process and providing the expected support to user
space applications.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>